move code to new file truncate_rows.py
